requirements.txt
audiocraft
azure-ai-vision
opencv-python
pydub
simpleaudio
openai
pillow